pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cogniflow"
version = "1.0.0"
description = "Assistente DSA con AI locale e interfaccia accessibile"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "DSA Developer", email = "your.email@example.com"},
]
maintainers = [
    {name = "DSA Developer", email = "your.email@example.com"},
]
keywords = ["DSA", "accessibility", "AI", "speech", "education"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyQt6>=6.5.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
]
ocr = [
    "pytesseract>=0.3.10",
]
speech = [
    "pyaudio>=0.2.13",
]

[project.scripts]
cogniflow = "assistente_dsa.main_00_launcher:main"

[tool.setuptools]
packages = ["assistente_dsa"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.qml", "*.json", "*.md", "*.png", "*.ico"]

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    "venv/",
    "env/",
    ".venv/",
]
ignore = [
    "E501",  # Line too long (handled by max-line-length)
    "W503",  # Line break before binary operator
    "E203",  # Whitespace before ':' (conflicts with black)
    "E402",  # Module level import not at top of file
    "F401",  # Imported but unused
]
per-file-ignores = [
    "tests/*.py:E501,E203,W503",
    "*/test_*.py:E501,E203,W503",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "PyQt6.*",
    "vosk.*",
    "pytesseract.*",
    "pyaudio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"